time: 8-28-00 1800 UTC changes: make_fnam_lp.f:             new FORTRAN jacket version of 'c_time2fname' that is more portable
         c_time2fname.c & Makefile:
            phased out of library

         lib/*.f
            Fixes for Linux warnings and added 'cdoc' documentation

         obs_driver.x:
            CWB mesonet improvements
            now logs counts of various types of pressure obs

         ingest_pro.exe:
            Linda's qc_flag changes for better Linux compatability

         lapsplot.x:
            added cloud image 'cg' using solid color fill
            added color for x-sect contours
            added colorfill for x-sect 'cf' cloud image plot
            fixed Linux compiler warnings and other cleanup

+++ src/ingest/profiler/ingest_sub_pro.f
@@ -121,7 +121,7 @@ C
 
         character*6 prof_name(n_profilers)
         character*9 a9_timeObs
-        integer timeObs
+        integer*4 timeObs
 
         real*4 lat(NX_L,NY_L),lon(NX_L,NY_L)
         real*4 topo(NX_L,NY_L)
